backend/requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.2.10
backcall==0.2.0
beautifulsoup4==4.9.1
billiard==3.3.0.23
bs4==0.0.1
celery==3.1.26.post2
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.9.2
decorator==4.4.2
Django==3.0.8
django-celery==3.3.1
django-mysql==3.7.1
django-redis==4.12.1
djangorestframework==3.11.0
future==0.18.2
idna==2.10
importlib-metadata==1.7.0
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
kombu==3.0.37
lxml==4.5.2
parso==0.7.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
pycparser==2.20
Pygments==2.6.1
PyMySQL==0.10.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
traitlets==4.3.3
urllib3==1.25.9
vine==1.3.0
wcwidth==0.2.5
zipp==3.1.0
